feat(--cpus): add support for --cpus flag
This commit adds support for the --cpus flag, allowing users to reproduce container CPU limits. The `docker inspect` command outputs this value as `NanoCpus`, an integer. However, the `docker run` command expects a floating-point number for the `--cpus` flag.

This implementation correctly parses the `NanoCpus` integer and converts it to a floating-point number. For example, a value of 1,000,000,000 NanoCpus will be correctly represented as 1.0 in the generated `docker run` command.